Bugfix: missing SqlColumnId must be comparable when using `partial` SQL Heuristic score.
